# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: DSCR Loan Automation Platform API
  description: |
    Comprehensive API for the DSCR (Debt Service Coverage Ratio) Loan Automation Platform.

    This platform automates the full lifecycle of DSCR refinance loans from lead intake
    through funding, with native integration to ICE Mortgage Technology Encompass LOS.

    ## Authentication
    All endpoints require Bearer token authentication via OAuth 2.0.

    ## Rate Limiting
    API calls are limited to 1000 requests per minute per client.

    ## Versioning
    The API is versioned via URL path (e.g., /api/v1/).
  version: 1.0.0
  contact:
    name: DSCR Platform Team
    email: support@dscrplatform.com
servers:
  - url: https://api.dscrplatform.com/api/v1
    description: Production
  - url: https://staging-api.dscrplatform.com/api/v1
    description: Staging
...
paths:
...
  /analytics/funnel:
    get:
      tags: [Analytics]
      summary: Get funnel conversion metrics
      operationId: getFunnelMetrics
      description: |
        Returns lead-to-funded conversion metrics with stage-by-stage breakdown.
        Useful for tracking overall funnel performance and identifying bottlenecks.
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date-time
          description: Start date for the period
        - name: to
          in: query
          schema:
            type: string
            format: date-time
          description: End date for the period
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [day, week, month]
          description: Time period grouping
        - name: loanOfficerId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by loan officer
      responses:
        '200':
          description: Funnel metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FunnelMetrics'
  /analytics/marketing:
    get:
      tags: [Analytics]
      summary: Get marketing attribution metrics
      operationId: getMarketingMetrics
      description: |
        Returns marketing performance metrics including source attribution,
        email campaign metrics, and website conversion rates.
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Marketing metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketingMetrics'
  /analytics/pipeline:
    get:
      tags: [Analytics]
      summary: Get pipeline metrics
      operationId: getPipelineMetrics
      description: |
        Returns real-time pipeline metrics including volume by milestone,
        aging analysis, and SLA breaches.
      parameters:
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [milestone, loanOfficer, processor]
          description: Grouping dimension
      responses:
        '200':
          description: Pipeline metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineMetrics'
...
  /analytics/velocity:
    get:
      tags: [Analytics]
      summary: Get velocity metrics
      operationId: getVelocityMetrics
      description: |
        Returns time-to-fund velocity metrics over time, useful for
        tracking operational efficiency trends.
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [day, week, month]
      responses:
        '200':
          description: Velocity metrics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VelocityMetrics'
  /analytics/events:
    post:
      tags: [Analytics]
      summary: Track analytics event
      operationId: trackEvent
      description: |
        Records a client-side analytics event for funnel tracking,
        page views, and conversion attribution.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyticsEvent'
      responses:
        '202':
          description: Event accepted for processing
components:
...
  schemas:
...
    FunnelStageMetrics:
      type: object
      properties:
        stage:
          type: string
          enum: [LEAD, CONTACTED, QUALIFIED, APPLICATION, PRE_APPROVED, FUNDED]
        count:
          type: integer
        conversionRate:
          type: number
          nullable: true
        previousStage:
          type: string
    FunnelMetrics:
      type: object
      properties:
        stages:
          type: array
          items:
            $ref: '#/components/schemas/FunnelStageMetrics'
        overallConversion:
          type: number
        period:
          type: object
          properties:
            from:
              type: string
              format: date-time
            to:
              type: string
              format: date-time
    MarketingSourceMetrics:
      type: object
      properties:
        source:
          type: string
        leads:
          type: integer
        conversions:
          type: integer
        conversionRate:
          type: number
        revenueCents:
          type: integer
    EmailMetrics:
      type: object
      properties:
        sent:
          type: integer
        opened:
          type: integer
        openRate:
          type: number
        clicked:
          type: integer
        ctr:
          type: number
        converted:
          type: integer
    WebMetrics:
      type: object
      properties:
        visitors:
          type: integer
        leadSubmissions:
          type: integer
        conversionRate:
          type: number
    MarketingMetrics:
      type: object
      properties:
        bySource:
          type: array
          items:
            $ref: '#/components/schemas/MarketingSourceMetrics'
        emailMetrics:
          $ref: '#/components/schemas/EmailMetrics'
        webMetrics:
          $ref: '#/components/schemas/WebMetrics'
        period:
          type: object
          properties:
            from:
              type: string
              format: date-time
            to:
              type: string
              format: date-time
    PipelineMilestoneMetrics:
      type: object
      properties:
        milestone:
          type: string
        count:
          type: integer
        volumeCents:
          type: integer
        avgDaysInStage:
          type: number
    SLABreach:
      type: object
      properties:
        applicationId:
          type: string
          format: uuid
        loanNumber:
          type: string
        milestone:
          type: string
        daysInStage:
          type: number
        slaHours:
          type: integer
        breachedAt:
          type: string
          format: date-time
    PipelineMetrics:
      type: object
      properties:
        byMilestone:
          type: array
          items:
            $ref: '#/components/schemas/PipelineMilestoneMetrics'
        slaBreaches:
          type: array
          items:
            $ref: '#/components/schemas/SLABreach'
        totalVolumeCents:
          type: integer
        totalCount:
          type: integer
...
    VelocityMetrics:
      type: object
      properties:
        period:
          type: string
        avgDaysLeadToFund:
          type: number
        count:
          type: integer
    AnalyticsEvent:
      type: object
      required: [event]
      properties:
        event:
          type: string
          description: Event name (e.g., page_view, lead_form_started)
        properties:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        sessionId:
          type: string
        userId:
          type: string
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []